fix(procurement): partial-receipt budget posting + commitment parse bug
Found via live HTTP smoke of the receiving flow:
- commitment fallback INSERT used $4 as both numeric amount and in "$4 > 0"
  (integer), so Postgres deduced inconsistent types and the approve PATCH
  failed at parse time → "Failed to update request". Cast $4::numeric.
- received value never reached Actual spend when the DB had no finance account
  (expenses.account_id is NOT NULL): the receipt route silently skipped the
  expense, so committed dropped but actual stayed 0. Auto-provision a default
  'Project Cash' account in the receipts + project-expense routes so received
  accepted quantity always posts as actual spend.

Verified: 51/51 DB tests, build clean (306 routes). Live smoke confirmed
expected-resources, partial receive→resource sync, and over-receipt block.

+++ src/app/api/projects/[id]/procurement/[prid]/route.js
@@ -75,9 +75,9 @@ export async function PATCH(request, { params }) {
         // Fallback: a request with no lines still reserves its header total.
         await query(
           `INSERT INTO commitments (project_id, work_item_id, procurement_request_id, amount, currency, status, budget_category, created_by)
-           SELECT $1::uuid,$2::uuid,$3::uuid,$4,$5,'open',COALESCE($6,'other'),$7::uuid
+           SELECT $1::uuid,$2::uuid,$3::uuid,$4::numeric,$5,'open',COALESCE($6,'other'),$7::uuid
            WHERE NOT EXISTS (SELECT 1 FROM commitments WHERE procurement_request_id=$3::uuid)
-             AND $4 > 0`,
+             AND $4::numeric > 0`,
           [id, pr.work_item_id || null, prid, pr.total_est_cost, pr.currency, pr.budget_category || null, auth.userId]);
 
         // Mark lines ordered and create EXPECTED project resources (one per line)
